+++ src/web/w-websrv.c
@@ -194,7 +194,7 @@ saiw_lp_state(void *userobj, void *sh, lws_ss_constate_t state,
 	saiw_websrv_t *m = (saiw_websrv_t *)userobj;
 	struct vhd *vhd = (struct vhd *)m->opaque_data;
 
-	lwsl_user("%s: %s, ord 0x%x\n", __func__, lws_ss_state_name(state),
+	lwsl_user("%s: %s, ord 0x%x\n", __func__, lws_ss_state_name((int)state),
 		  (unsigned int)ack);
 
 	switch (state) {
